Refactor metrics SDK paths
Fixes #2666
