+++ docs/conf.py
@@ -97,6 +97,14 @@ nitpicky = True
 # https://github.com/sphinx-doc/sphinx/pull/3744
 nitpick_ignore = [
     ("py:class", "ValueT"),
+    (
+        "py:class",
+        "opentelemetry.sdk._metrics._internal.instrument._Synchronous",
+    ),
+    (
+        "py:class",
+        "opentelemetry.sdk._metrics._internal.instrument._Asynchronous",
+    ),
     # Even if wrapt is added to intersphinx_mapping, sphinx keeps failing
     # with "class reference target not found: ObjectProxy".
     ("py:class", "ObjectProxy"),
