--- docs/sdk/metrics.aggregation.rst
@@ -1,7 +1,0 @@
-opentelemetry.sdk._metrics.aggregation
-======================================
-
-.. automodule:: opentelemetry.sdk._metrics.aggregation
-    :members:
-    :undoc-members:
-    :no-show-inheritance:
