--- docs/sdk/metrics.metric_reader.rst
@@ -1,7 +1,0 @@
-opentelemetry.sdk._metrics.metric_reader
-==========================================
-
-.. automodule:: opentelemetry.sdk._metrics.metric_reader
-    :members:
-    :undoc-members:
-    :show-inheritance:
