--- docs/sdk/metrics.point.rst
@@ -1,7 +1,0 @@
-opentelemetry.sdk._metrics.point
-================================
-
-.. automodule:: opentelemetry.sdk._metrics.point
-    :members:
-    :undoc-members:
-    :no-show-inheritance:
