+++ docs/sdk/metrics.rst
@@ -13,10 +13,6 @@ Submodules
 
 .. toctree::
 
-   metrics.view
-   metrics.aggregation
-   metrics.metric_reader
-   metrics.point
    metrics.export
 
 .. automodule:: opentelemetry.sdk._metrics
