--- docs/sdk/metrics.view.rst
@@ -1,7 +1,0 @@
-opentelemetry.sdk._metrics.view
-==========================================
-
-.. automodule:: opentelemetry.sdk._metrics.view
-    :members:
-    :undoc-members:
-    :show-inheritance:
