+++ exporter/opentelemetry-exporter-prometheus/src/opentelemetry/exporter/prometheus/__init__.py
@@ -77,8 +77,13 @@ from prometheus_client.core import (
 )
 from prometheus_client.core import Metric as PrometheusMetric
 
-from opentelemetry.sdk._metrics.metric_reader import MetricReader
-from opentelemetry.sdk._metrics.point import Gauge, Histogram, Metric, Sum
+from opentelemetry.sdk._metrics.export import (
+    Gauge,
+    Histogram,
+    Metric,
+    MetricReader,
+    Sum,
+)
 
 _logger = getLogger(__name__)
