+++ exporter/opentelemetry-exporter-prometheus/tests/test_prometheus_exporter.py
@@ -23,7 +23,8 @@ from opentelemetry.exporter.prometheus import (
     PrometheusMetricReader,
     _CustomCollector,
 )
-from opentelemetry.sdk._metrics.point import AggregationTemporality, Histogram
+from opentelemetry.sdk._metrics import AggregationTemporality
+from opentelemetry.sdk._metrics.export import Histogram
 from opentelemetry.test.metrictestutil import (
     _generate_gauge,
     _generate_metric,
