+++ opentelemetry-sdk/src/opentelemetry/sdk/_metrics/__init__.py
@@ -18,6 +18,24 @@ from opentelemetry.sdk._metrics._internal import (  # noqa: F401
     Meter,
     MeterProvider,
 )
+from opentelemetry.sdk._metrics._internal.aggregation import (  # noqa: F401
+    Aggregation,
+    AggregationTemporality,
+    DefaultAggregation,
+    DropAggregation,
+    ExplicitBucketHistogramAggregation,
+    LastValueAggregation,
+    SumAggregation,
+)
+from opentelemetry.sdk._metrics._internal.instrument import (  # noqa: F401
+    Counter,
+    Histogram,
+    ObservableCounter,
+    ObservableGauge,
+    ObservableUpDownCounter,
+    UpDownCounter,
+)
+from opentelemetry.sdk._metrics._internal.view import View  # noqa: F401
 
 __all__ = []
 for key, value in globals().copy().items():
