+++ opentelemetry-sdk/src/opentelemetry/sdk/_metrics/_internal/measurement_consumer.py
@@ -12,23 +12,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# pylint: disable=unused-import
+
 from abc import ABC, abstractmethod
 from threading import Lock
-from typing import TYPE_CHECKING, Dict, Iterable, List, Mapping
+from typing import Dict, Iterable, List, Mapping
 
+import opentelemetry.sdk._metrics
+import opentelemetry.sdk._metrics._internal.instrument
+import opentelemetry.sdk._metrics._internal.sdk_configuration
 from opentelemetry._metrics._internal.instrument import CallbackOptions
+from opentelemetry.sdk._metrics._internal.aggregation import (
+    AggregationTemporality,
+)
+from opentelemetry.sdk._metrics._internal.measurement import Measurement
 from opentelemetry.sdk._metrics._internal.metric_reader_storage import (
     MetricReaderStorage,
 )
-from opentelemetry.sdk._metrics._internal.sdk_configuration import (
-    SdkConfiguration,
-)
-from opentelemetry.sdk._metrics.measurement import Measurement
-from opentelemetry.sdk._metrics.metric_reader import MetricReader
-from opentelemetry.sdk._metrics.point import AggregationTemporality, Metric
-
-if TYPE_CHECKING:
-    from opentelemetry.sdk._metrics._internal.instrument import _Asynchronous
+from opentelemetry.sdk._metrics.export import Metric
 
 
 class MeasurementConsumer(ABC):
@@ -37,44 +38,59 @@ class MeasurementConsumer(ABC):
         pass
 
     @abstractmethod
-    def register_asynchronous_instrument(self, instrument: "_Asynchronous"):
+    def register_asynchronous_instrument(
+        self,
+        instrument: (
+            "opentelemetry.sdk._metrics._internal.instrument_Asynchronous"
+        ),
+    ):
         pass
 
     @abstractmethod
     def collect(
         self,
-        metric_reader: MetricReader,
+        metric_reader: "opentelemetry.sdk._metrics.MetricReader",
         instrument_type_temporality: Dict[type, AggregationTemporality],
     ) -> Iterable[Metric]:
         pass
 
 
 class SynchronousMeasurementConsumer(MeasurementConsumer):
-    def __init__(self, sdk_config: SdkConfiguration) -> None:
+    def __init__(
+        self,
+        sdk_config: "opentelemetry.sdk._metrics._internal.SdkConfiguration",
+    ) -> None:
         self._lock = Lock()
         self._sdk_config = sdk_config
         # should never be mutated
-        self._reader_storages: Mapping[MetricReader, MetricReaderStorage] = {
+        self._reader_storages: Mapping[
+            "opentelemetry.sdk._metrics.MetricReader", MetricReaderStorage
+        ] = {
             reader: MetricReaderStorage(
                 sdk_config, reader._instrument_class_aggregation
             )
             for reader in sdk_config.metric_readers
         }
-        self._async_instruments: List["_Asynchronous"] = []
+        self._async_instruments: List[
+            "opentelemetry.sdk._metrics._internal.instrument._Asynchronous"
+        ] = []
 
     def consume_measurement(self, measurement: Measurement) -> None:
         for reader_storage in self._reader_storages.values():
             reader_storage.consume_measurement(measurement)
 
     def register_asynchronous_instrument(
-        self, instrument: "_Asynchronous"
+        self,
+        instrument: (
+            "opentelemetry.sdk._metrics._internal.instrument._Asynchronous"
+        ),
     ) -> None:
         with self._lock:
             self._async_instruments.append(instrument)
 
     def collect(
         self,
-        metric_reader: MetricReader,
+        metric_reader: "opentelemetry.sdk._metrics.MetricReader",
         instrument_type_temporality: Dict[type, AggregationTemporality],
     ) -> Iterable[Metric]:
         with self._lock:
