+++ opentelemetry-sdk/src/opentelemetry/sdk/_metrics/_internal/sdk_configuration.py
@@ -1,15 +1,28 @@
+# Copyright The OpenTelemetry Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# pylint: disable=unused-import
+
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Sequence
+from typing import Sequence
 
-from opentelemetry.sdk._metrics.metric_reader import MetricReader
-from opentelemetry.sdk.resources import Resource
-
-if TYPE_CHECKING:
-    from opentelemetry.sdk._metrics.view import View
+import opentelemetry.sdk._metrics
+import opentelemetry.sdk.resources
 
 
 @dataclass
 class SdkConfiguration:
-    resource: Resource
-    metric_readers: Sequence[MetricReader]
-    views: Sequence["View"]
+    resource: "opentelemetry.sdk.resources.Resource"
+    metric_readers: Sequence["opentelemetry.sdk._metrics.MetricReader"]
+    views: Sequence["opentelemetry.sdk._metrics.View"]
