+++ opentelemetry-sdk/tests/metrics/integration_test/test_disable_default_views.py
@@ -14,10 +14,8 @@
 
 from unittest import TestCase
 
-from opentelemetry.sdk._metrics import MeterProvider
-from opentelemetry.sdk._metrics.aggregation import DropAggregation
+from opentelemetry.sdk._metrics import DropAggregation, MeterProvider, View
 from opentelemetry.sdk._metrics.export import InMemoryMetricReader
-from opentelemetry.sdk._metrics.view import View
 
 
 class TestDisableDefaultViews(TestCase):
