+++ opentelemetry-sdk/tests/metrics/test_measurement_consumer.py
@@ -15,6 +15,7 @@
 from unittest import TestCase
 from unittest.mock import MagicMock, Mock, patch
 
+from opentelemetry.sdk._metrics import AggregationTemporality
 from opentelemetry.sdk._metrics._internal.measurement_consumer import (
     MeasurementConsumer,
     SynchronousMeasurementConsumer,
@@ -22,7 +23,6 @@ from opentelemetry.sdk._metrics._internal.measurement_consumer import (
 from opentelemetry.sdk._metrics._internal.sdk_configuration import (
     SdkConfiguration,
 )
-from opentelemetry.sdk._metrics.point import AggregationTemporality
 
 
 @patch(
