+++ opentelemetry-sdk/tests/metrics/test_view_instrument_match.py
@@ -15,6 +15,14 @@
 from unittest import TestCase
 from unittest.mock import MagicMock, Mock
 
+from opentelemetry.sdk._metrics import (
+    AggregationTemporality,
+    Counter,
+    DefaultAggregation,
+    DropAggregation,
+    LastValueAggregation,
+    View,
+)
 from opentelemetry.sdk._metrics._internal._view_instrument_match import (
     _ViewInstrumentMatch,
 )
@@ -22,18 +30,11 @@ from opentelemetry.sdk._metrics._internal.aggregation import (
     _DropAggregation,
     _LastValueAggregation,
 )
+from opentelemetry.sdk._metrics._internal.measurement import Measurement
 from opentelemetry.sdk._metrics._internal.sdk_configuration import (
     SdkConfiguration,
 )
-from opentelemetry.sdk._metrics.aggregation import (
-    DefaultAggregation,
-    DropAggregation,
-    LastValueAggregation,
-)
-from opentelemetry.sdk._metrics.instrument import Counter
-from opentelemetry.sdk._metrics.measurement import Measurement
-from opentelemetry.sdk._metrics.point import AggregationTemporality, Metric
-from opentelemetry.sdk._metrics.view import View
+from opentelemetry.sdk._metrics.export import Metric
 
 
 class Test_ViewInstrumentMatch(TestCase):
